+++ wp-php-console.php
@@ -3,12 +3,13 @@
  * @link              https://github.com/nekojira/wp-php-console/
  * @since             1.0.0
  * @package           WP_PHP_Console
+ * @author            nekojira <fulvio@nekojira.com>
  *
  * @wordpress-plugin
  * Plugin Name:       WP PHP Console
  * Plugin URI:        https://github.com/nekojira/wp-php-console/
  * Description:       An implementation of PHP Console for WordPress. Easily debug and trace PHP errors and warnings from your Chrome dev tools console using a Google Chrome extension.
- * Version:           1.0.0
+ * Version:           1.1.0
  * Author:            nekojira
  * Author URI:        https://github.com/nekojira/
  * License:           GPL-2.0+
@@ -19,7 +20,19 @@
 
 if ( ! defined( 'WPINC' ) ) die;
 
-require_once plugin_dir_path( __FILE__ ) . 'lib/vendor/autoload.php';
+/**
+ * Include PhpConsole server library.
+ * @link https://github.com/barbushin/php-console
+ * Copyright (c) 2011-2013 by Barbushin Sergey <barbushin@gmail.com>.
+ */
+require_once plugin_dir_path( __FILE__ ) . 'lib/php-console/src/PhpConsole/__autoload.php';
+
+/**
+ * The main class of this plugin.
+ */
 require_once plugin_dir_path( __FILE__ ) . 'lib/class-wp-php-console.php';
 
+/**
+ * Instantiate this plugin.
+ */
 $WP_PHP_Console = new WP_PHP_Console();