+++ ChangeLog.md
@@ -1,3 +1,8 @@
-### 1.00 (07 nov 2014)
+### 1.1.0 (07 nov 2014)
+* Added donation link/button.
+* PHP Console server is now instantiated later, allowing to catch all your theme functions too.
+* Included PHP Console server library as git submodule rather than a composer dependency.
+
+### 1.0.0 (06 nov 2014)
 * Added three options to set a custom password, enable on SSL only, authorized IP ranges.
 * First public release.